+++ src/test/java/integration/fakeevents/BaseIntegrationTestWithFakeEvents.java
@@ -1,5 +1,6 @@
-package integration;
+package integration.fakeevents;
 
+import integration.BaseIntegrationTest;
 import org.springframework.cloud.contract.wiremock.AutoConfigureWireMock;
 
 @AutoConfigureWireMock(port = 8092)
